requirements.txt

flask
flask-cors
requests
google-genai
supabase
pytz
breeze-connect
